spoof-detector/requirements.txt
paho-mqtt<2.0.0
requests
geopy
schedule
influxdb
